#23: Move EnvironmentEndpoints,ProjectEndpoints,LinkEndpoints,AutomationEndpoints,HealthcheckEndpoints to endpoints.py
